update
git-svn-id: https://svn.r-project.org/R-dev-web/trunk@2956 c52295ea-58df-0310-926a-d16021944841

+++ CRAN/QA/BDR/gannet/tests-48x/unpack.sh
@@ -1,13 +1,13 @@
 for p in ../contrib/*.tar.gz; do
   pkgname=`basename $p | sed  -e 's/_.*//'`
-if test $p -nt $pkgname.in; then
+if test $pkgname != TCC -a $p -nt $pkgname.in; then
   echo $pkgname
   rm -rf $pkgname
   tar zxf $p
   touch -r $p $pkgname.in
 fi
 done
-for p in ../3.1.0/Recommended/MASS*.tar.gz; do
+for p in ../3.1.0/Other/*.tar.gz; do
   pkgname=`basename $p | sed  -e 's/_.*//'`
 if test $p -nt $pkgname.in; then
   echo $pkgname
@@ -16,4 +16,3 @@ if test $p -nt $pkgname.in; then
   touch -r $p $pkgname.in
 fi
 done
-
